Refactor SPI example to facilitate adding more types of examples.
