+++ examples/shardingsphere-jdbc-example/single-feature-example/extension-example/custom-sharding-algortihm-example/spi-based-sharding-algorithm-example/spi-based-sharding-raw-jdbc-example/src/main/resources/META-INF/services/org.apache.shardingsphere.sharding.spi.ShardingAlgorithm
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
  
-org.apache.shardingsphere.example.extension.sharding.algortihm.spibased.spi.SpiBasedShardingAlgorithm
+org.apache.shardingsphere.example.extension.spibased.sharding.raw.jdbc.fixture.SPIBasedStandardShardingAlgorithmFixture
